Create Task-Manager
Task Manager - Day 35 Challenge of Replit's 100 Days of Code. In review, I was able to eliminate some instances of user error causing the program to revert to the top menu without performing the desired task. Also changed my list print function to require the user to manually return to the top menu, as opposed to a timed clear and return.