Create mokeDex.py
Day 46 challenge from Replit. This program is really basic, but made for a good intro to 2D dictionaries. 
I could add more fluff to indicate to the user what is happening or has/hasn't happened, but that's not where I need practice. Adding formatting to the prettyPrint function is the likely where I could use some additional practice (alignment, color code based on element).
Defining functions at the top and simplifying the flow of the program make this a lot easier to look at.